// users-service/src/domain/model/custPrimaryDetails.interface.ts
import { ClientStatus } from '../enum/clientStatus.enum';
import { ClientType } from '../enum/clientType.enum';
import { Gender } from '../enum/gender.enum';
import { IdCardStatus } from '../enum/id-card-status.enum';
import { IBase } from './base.interface';

export interface ICustPrimaryDetails extends IBase {
  id: string;
  leadId: string;
  cognitoId: string;
  clientType: ClientType;
  clientStatus: ClientStatus;
  msisdnCountryCode: string;
  msisdn: string;
  nationalIdNumber: string;
  email: string;
  surname: string;
  givenName: string;
  preferredName: string;
  nationality: string;
  gender: Gender;
  dateOfBirth: Date;
  NINOCR: string;
  cardNumber: string;
  dateOfExpiry: Date;
  scannedImageFront: string;
  scannedImageBack: string;
  selfieImage: string;
  faceMatchPercentage: number;
  faceMatchStatus: string;
  liveliessCheckPercenatge: number;
  livelinessCheckStatus: string;
  totalLoans: number;
  creditExpiryTime: string;
  idExpiryDays: number;
  idStatus: IdCardStatus;
  smsNextHours: number;
  optOutReason: string;
  optOutFeedback: string;
  optOutDate: Date;
  availableCreditLimit: number;
}